+++ docs/identifier_lookup/CHANGELOG.md
@@ -3,7 +3,8 @@
 All notable changes to this gear are documented in this file.
 
 ## 1.3.1
-
+* Make `form_configs_file` only required for `nacc` direction
+* For the `center` direction (reverse lookup), adds sleep of 1 second every 1000 records to reduce connectivity issues
 * Rebuilt for new preprocessing checks in config
 
 ## 1.3.0